Add price column to sample data

+++ bookstore/bookstore_db.py
@@ -36,10 +36,13 @@ def init_db():
     db = get_db()
 
     # Cleanup database
-    db["books"].delete_many({})
-    db["inventory"].delete_many({})
-    db["customers"].delete_many({})
-    db["orders"].delete_many({})
+    clean_up = False
+
+    if clean_up:
+        db["books"].delete_many({})
+        db["inventory"].delete_many({})
+        db["customers"].delete_many({})
+        db["orders"].delete_many({})
 
     # Insert sample data for books
     db.books.insert_many(
@@ -49,6 +52,7 @@ def init_db():
             "author": ["Michelle Obama"],
             "language": "English",
             "isbn": "1524763136",
+            "price": 10.5,
             "published_date": "2018-11-13",
             "publisher": "Crown Publishing Group"
         }, {
@@ -57,6 +61,7 @@ def init_db():
             "author": ["David Baldacci"],
             "language": "English",
             "isbn": "1538761459",
+            "price": 15.5,
             "published_date": "2019-10-08",
             "publisher": "Grand Central Publishing"
         }, {
@@ -65,6 +70,7 @@ def init_db():
             "author": ["Anna Burns"],
             "language": "English",
             "isbn": "1644450003",
+            "price": 25.5,
             "published_date": "2018-12-04",
             "publisher": "Graywolf Press"
         }]
